requirements.txt
cn_clip>=1.5.1
torch>=1.8.0
torchvision>=0.9.0
faiss-cpu>=1.7.0
streamlit>=1.28.0
pandas>=1.5.0
pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
tqdm>=4.64.0
matplotlib>=3.5.0
plotly>=5.0.0
loguru>=0.7.0 